Move stuff from Providers folder to apropriate features
