+++ ArmaForces.ArmaServerManager.Tests/Providers/Server/ServerProviderTests.cs
@@ -5,8 +5,8 @@ using ArmaForces.Arma.Server.Features.Modsets;
 using ArmaForces.Arma.Server.Features.Processes;
 using ArmaForces.Arma.Server.Features.Servers;
 using ArmaForces.Arma.Server.Tests.Helpers;
-using ArmaForces.ArmaServerManager.Providers;
-using ArmaForces.ArmaServerManager.Providers.Server;
+using ArmaForces.ArmaServerManager.Features.Modsets;
+using ArmaForces.ArmaServerManager.Features.Servers;
 using AutoFixture;
 using AutoFixture.Kernel;
 using CSharpFunctionalExtensions;
