+++ ArmaForces.ArmaServerManager/Api/Servers/ServersController.cs
@@ -6,8 +6,8 @@ using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Features.Servers.DTOs;
 using ArmaForces.ArmaServerManager.Api.Servers.DTOs;
 using ArmaForces.ArmaServerManager.Features.Jobs;
+using ArmaForces.ArmaServerManager.Features.Servers;
 using ArmaForces.ArmaServerManager.Infrastructure.Authentication;
-using ArmaForces.ArmaServerManager.Providers.Server;
 using ArmaForces.ArmaServerManager.Services;
 using CSharpFunctionalExtensions;
 using Microsoft.AspNetCore.Http;
