+++ ArmaForces.ArmaServerManager/Features/Mods/DependencyInjection/ModsServiceCollectionExtensions.cs
@@ -1,7 +1,7 @@
-﻿using ArmaForces.ArmaServerManager.Features.Modsets.DependencyInjection;
+﻿using ArmaForces.ArmaServerManager.Features.Modsets;
+using ArmaForces.ArmaServerManager.Features.Modsets.DependencyInjection;
 using ArmaForces.ArmaServerManager.Features.Steam;
 using ArmaForces.ArmaServerManager.Features.Steam.Content;
-using ArmaForces.ArmaServerManager.Providers;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace ArmaForces.ArmaServerManager.Features.Mods.DependencyInjection
