+++ ArmaForces.ArmaServerManager/Features/Modsets/IModsetProvider.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Features.Modsets;
 using CSharpFunctionalExtensions;
 
-namespace ArmaForces.ArmaServerManager.Providers
+namespace ArmaForces.ArmaServerManager.Features.Modsets
 {
     public interface IModsetProvider
     {
