+++ ArmaForces.ArmaServerManager/Features/Status/StatusProvider.cs
@@ -6,8 +6,8 @@ using ArmaForces.Arma.Server.Features.Servers.DTOs;
 using ArmaForces.ArmaServerManager.Features.Jobs;
 using ArmaForces.ArmaServerManager.Features.Jobs.Extensions;
 using ArmaForces.ArmaServerManager.Features.Jobs.Persistence.Models;
+using ArmaForces.ArmaServerManager.Features.Servers;
 using ArmaForces.ArmaServerManager.Features.Status.Models;
-using ArmaForces.ArmaServerManager.Providers.Server;
 using ArmaForces.ArmaServerManager.Services;
 using CSharpFunctionalExtensions;
