+++ ArmaForces.ArmaServerManager/Services/ModsVerificationService.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Features.Modsets;
 using ArmaForces.ArmaServerManager.Features.Mods;
-using ArmaForces.ArmaServerManager.Providers;
+using ArmaForces.ArmaServerManager.Features.Modsets;
 using CSharpFunctionalExtensions;
 
 namespace ArmaForces.ArmaServerManager.Services
