+++ ArmaForces.ArmaServerManager/Services/ServerStartupService.cs
@@ -5,8 +5,8 @@ using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Features.Modsets;
 using ArmaForces.ArmaServerManager.Features.Missions;
 using ArmaForces.ArmaServerManager.Features.Missions.DTOs;
-using ArmaForces.ArmaServerManager.Providers;
-using ArmaForces.ArmaServerManager.Providers.Server;
+using ArmaForces.ArmaServerManager.Features.Modsets;
+using ArmaForces.ArmaServerManager.Features.Servers;
 using CSharpFunctionalExtensions;
 
 namespace ArmaForces.ArmaServerManager.Services
