+++ ArmaForces.ArmaServerManager/Services/StartupService.cs
@@ -1,8 +1,8 @@
-using System;
+﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
 using ArmaForces.ArmaServerManager.Features.Jobs;
-using ArmaForces.ArmaServerManager.Providers.Server;
+using ArmaForces.ArmaServerManager.Features.Servers;
 using Hangfire;
 using Microsoft.Extensions.Hosting;
